Enable the optional addition of serving service ca to token secrets
Previously the operand always added the serving service ca to token
secrets. This commit defaults that behavior to not occur but allows it
be enabled by setting an UnsupportedConfigOverride named
enableDeprecatedAndRemovedServiceCAKeyUntilNextRelease_ThisMakesClusterImpossibleToUpgrade. When
the override is set to true, the operator will be marked as not
upgradeable. This is a temporary measure to allow but discourage the
behavior prior to its removal in 4.6.