Bug 1806640: fix potential errors in Prometheus alerts
There is no need to use the 'on' keyword to compare
kube_poddisruptionbudget_status_expected_pods and
kube_poddisruptionbudget_status_desired_healthy because both series have
the same label set. In addition when there are 2 instances of
kube-state-metrics running, it triggers evaluation errors in Prometheus.

To ensure that the expression returns one item per (namespace,
poddisruptionbudget) tuple, it is wrapped by the 'max' aggregation
operator.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>