BUG 1853889: Change port for kube-controller-manager-recovery-controller
PR[1] added logic for checking port availability in recovery-controller.
PR[2] change the port of HAProxy to 9443, due to yet another port conflict.
Both of the PRs cause the kube-controller-manager-recovery-controller to crash loop
on on-prem platforms such as oVirt, baremetal and openstack.
This PR changes the port ov the recovery controller.

[1] https://github.com/openshift/cluster-kube-controller-manager-operator/pull/421
[2] https://github.com/openshift/baremetal-runtimecfg/pull/59

Signed-off-by: Gal-Zaidman <gzaidman@redhat.com>