support custom CA bundle for AWS C2S
Communication with the AWS API in the C2S Secret region requires a
custom CA bundle. This bundle is stored in the "ca-bundle.pem" entry
in the cloud-config. The AWS_CA_BUNDLE environment variable informs
the AWS SDK to use a custom CA bundle and to specify the location of
the bundle.