manifests/0000_90_kube-controller-manager-operator_05_alerts: Template console links in alert descriptions
Prometheus alerts support Go templating [1], and this commit uses that
to provide more context like "which namespace?", "which
PodDisruptionBudget?", "where can I find that PDB in the in-cluster
web console?", and "what 'oc' command would I run to see
garbage-collection sync logs?".  This should make understanding the
context of the alert more straightforward, with the responder having
to dip into labels and guess.

Using |- for trimmed, block style strings avoids YAML parsers choking
on the "for more details: ..." colon with "mapping values are not
allowed in this context" and similar.

[1]: https://prometheus.io/docs/prometheus/latest/configuration/template_reference/