Throttle resourcesynccontroller on pairs with dual writers.
Writes from the instance of resourcesynccontroller in the operator can contend with writes from the
recovery sidecar's instance of the same controller. This is a preexisting issue, and the controller
doesn't have an effective way to limit resync frequency. Adding some fixed latency to each sync in
this way sets an upper bound on the churn rate when the two instances get into a feedback cycle.

The separate issues that can create a feedback cycle (e.g. the resourcesynccontroller degraded
operator status condition is shared by both instances) should be separately addressed "soon".